+++ sandbox/sb_image.cpp
@@ -4,18 +4,18 @@
 namespace Sandbox {
     
     bool Image::CheckBit(int x,int y,Resources* resources) {
-        x+=m_hotspot.x*m_width/m_src_w;
+        x+=int(m_hotspot.x*m_width/m_src_w);
         if (x<0)
             return false;
-        y+=m_hotspot.y*m_height/m_src_h;
+        y+=int(m_hotspot.y*m_height/m_src_h);
         if (y<0)
             return false;
         if (x>m_width)
             return false;
         if (y>m_height)
             return false;
-        x = m_src_x + x * m_src_w / m_width;
-        y = m_src_y + y * m_src_h / m_height;
+        x = int(m_src_x + x * m_src_w / m_width);
+        y = int(m_src_y + y * m_src_h / m_height);
         if (!m_texture) return false;
         BitmaskPtr bm = m_texture->GetBitmask(resources);
         if (!bm)
